reduce e2e complexity
