+++ test/e2e/e2e_test.go
@@ -12,101 +12,53 @@ import (
 	"testing"
 
 	"github.com/openshift/hypershift/cmd/version"
-	e2elog "github.com/openshift/hypershift/test/e2e/log"
-	"github.com/openshift/hypershift/test/e2e/scenarios"
 	"k8s.io/apimachinery/pkg/util/errors"
+
+	"github.com/bombsimon/logrusr"
+	"github.com/sirupsen/logrus"
 )
 
 var (
 	// opts are global options for the test suite bound in TestMain.
-	opts = &options{}
+	globalOpts = &options{}
 
-	// ctx should be used as the parent context for any test code, and will
+	// testContext should be used as the parent context for any test code, and will
 	// be cancelled if a SIGINT or SIGTERM is received. It's set up in TestMain.
-	ctx context.Context
+	testContext context.Context
 
-	log = e2elog.Logger
+	log = logrusr.NewLogger(logrus.New())
 )
-
-// TestScenarios runs all the e2e tests. Any new tests need to be added to this
-// list in order for them to run.
-func TestScenarios(t *testing.T) {
-	tests := map[string]func(t *testing.T){
-		"CreateCluster": scenarios.TestCreateCluster(ctx,
-			scenarios.TestCreateClusterOptions{
-				AWSCredentialsFile: opts.AWSCredentialsFile,
-				AWSRegion:          opts.Region,
-				PullSecretFile:     opts.PullSecretFile,
-				ReleaseImage:       opts.LatestReleaseImage,
-				ArtifactDir:        opts.ArtifactDir,
-				BaseDomain:         opts.BaseDomain,
-			}),
-		"UpgradeControlPlane": scenarios.TestUpgradeControlPlane(ctx,
-			scenarios.TestUpgradeControlPlaneOptions{
-				AWSCredentialsFile: opts.AWSCredentialsFile,
-				AWSRegion:          opts.Region,
-				BaseDomain:         opts.BaseDomain,
-				PullSecretFile:     opts.PullSecretFile,
-				FromReleaseImage:   opts.PreviousReleaseImage,
-				ToReleaseImage:     opts.LatestReleaseImage,
-				ArtifactDir:        opts.ArtifactDir,
-				Enabled:            opts.UpgradeTestsEnabled,
-			}),
-		"Autoscaling": scenarios.TestAutoscaling(ctx,
-			scenarios.TestAutoscalingOptions{
-				AWSCredentialsFile: opts.AWSCredentialsFile,
-				AWSRegion:          opts.Region,
-				PullSecretFile:     opts.PullSecretFile,
-				ReleaseImage:       opts.LatestReleaseImage,
-				ArtifactDir:        opts.ArtifactDir,
-				BaseDomain:         opts.BaseDomain,
-			}),
-		"AutoRepair": scenarios.TestAutoRepair(ctx,
-			scenarios.TestAutoRepairOptions{
-				AWSCredentialsFile: opts.AWSCredentialsFile,
-				AWSRegion:          opts.Region,
-				PullSecretFile:     opts.PullSecretFile,
-				ReleaseImage:       opts.LatestReleaseImage,
-				ArtifactDir:        opts.ArtifactDir,
-				BaseDomain:         opts.BaseDomain,
-			}),
-	}
-
-	for name := range tests {
-		fn := tests[name]
-		t.Run(name, fn)
-	}
-}
 
 // TestMain deals with global options and setting up a signal-bound context
 // for all tests to use.
 func TestMain(m *testing.M) {
-	flag.StringVar(&opts.AWSCredentialsFile, "e2e.aws-credentials-file", "", "path to AWS credentials")
-	flag.StringVar(&opts.Region, "e2e.aws-region", "us-east-1", "AWS region for clusters")
-	flag.StringVar(&opts.PullSecretFile, "e2e.pull-secret-file", "", "path to pull secret")
-	flag.StringVar(&opts.LatestReleaseImage, "e2e.latest-release-image", "", "The latest OCP release image for use by tests")
-	flag.StringVar(&opts.PreviousReleaseImage, "e2e.previous-release-image", "", "The previous OCP release image relative to the latest")
-	flag.StringVar(&opts.ArtifactDir, "e2e.artifact-dir", "", "The directory where cluster resources and logs should be dumped. If empty, nothing is dumped")
-	flag.StringVar(&opts.BaseDomain, "e2e.base-domain", "", "The ingress base domain for the cluster")
-	flag.BoolVar(&opts.UpgradeTestsEnabled, "e2e.upgrade-tests-enabled", false, "Enables upgrade tests")
-	flag.StringVar(&opts.ControlPlaneOperatorImage, "e2e.control-plane-operator-image", "", "The image to use for the control plane operator. If none specified, the default is used.")
+	flag.StringVar(&globalOpts.AWSCredentialsFile, "e2e.aws-credentials-file", "", "path to AWS credentials")
+	flag.StringVar(&globalOpts.Region, "e2e.aws-region", "us-east-1", "AWS region for clusters")
+	flag.StringVar(&globalOpts.PullSecretFile, "e2e.pull-secret-file", "", "path to pull secret")
+	flag.StringVar(&globalOpts.LatestReleaseImage, "e2e.latest-release-image", "", "The latest OCP release image for use by tests")
+	flag.StringVar(&globalOpts.PreviousReleaseImage, "e2e.previous-release-image", "", "The previous OCP release image relative to the latest")
+	flag.StringVar(&globalOpts.ArtifactDir, "e2e.artifact-dir", "", "The directory where cluster resources and logs should be dumped. If empty, nothing is dumped")
+	flag.StringVar(&globalOpts.BaseDomain, "e2e.base-domain", "", "The ingress base domain for the cluster")
+	flag.BoolVar(&globalOpts.UpgradeTestsEnabled, "e2e.upgrade-tests-enabled", false, "Enables upgrade tests")
+	flag.StringVar(&globalOpts.ControlPlaneOperatorImage, "e2e.control-plane-operator-image", "", "The image to use for the control plane operator. If none specified, the default is used.")
 
 	flag.Parse()
 
 	// Set defaults for the test options
-	if err := opts.Complete(); err != nil {
+	if err := globalOpts.Complete(); err != nil {
 		log.Error(err, "failed to set up global test options")
 		os.Exit(1)
 	}
 
 	// Validate the test options
-	if err := opts.Validate(); err != nil {
+	if err := globalOpts.Validate(); err != nil {
 		log.Error(err, "invalid global test options")
 		os.Exit(1)
 	}
 
 	// Set up a root context for all tests and set up signal handling
-	rootCtx, cancel := context.WithCancel(context.Background())
+	var cancel context.CancelFunc
+	testContext, cancel = context.WithCancel(context.Background())
 	sigs := make(chan os.Signal, 1)
 	signal.Notify(sigs, syscall.SIGINT, syscall.SIGTERM)
 	go func() {
@@ -114,10 +66,9 @@ func TestMain(m *testing.M) {
 		log.Info("tests received shutdown signal and will be cancelled")
 		cancel()
 	}()
-	ctx = rootCtx
 
 	// Everything's okay to run tests
-	log.Info("executing e2e tests", "options", opts)
+	log.Info("executing e2e tests", "options", globalOpts)
 	os.Exit(m.Run())
 }
