--- test/e2e/log/log.go
@@ -1,8 +1,0 @@
-package log
-
-import (
-	"github.com/bombsimon/logrusr"
-	"github.com/sirupsen/logrus"
-)
-
-var Logger = logrusr.NewLogger(logrus.New())
