--- test/e2e/util/log.go
@@ -1,5 +1,0 @@
-package util
-
-import e2elog "github.com/openshift/hypershift/test/e2e/log"
-
-var log = e2elog.Logger
